[ENGG-2083]: A/B exp setup for new rules empty state (#1963)
* refactor: getting started for A/B exp

* fix: create your first rule click

* fix: create your first rule

* cleanup

* fix: creat rule redirect

* fix: create rule redirect

* ab setup

* cleanup

* fix
