+++ app/src/features/rules/screens/rulesList/components/RulesList/components/RuleSelectionListDrawer/RuleSelectionListDrawer.tsx
@@ -1,6 +1,8 @@
 import React from "react";
 import { Drawer, DrawerProps } from "antd";
 import { RuleSelectionList, RuleSelectionListProps } from "../RuleSelectionList/RuleSelectionList";
+import { useIsRedirectFromCreateRulesRoute } from "../../hooks/useIsRedirectFromCreateRulesRoute";
+import { useLocation, useNavigate } from "react-router-dom";
 import "./RuleSelectionListDrawer.scss";
 
 interface RuleSelectionListDrawerProps extends Omit<RuleSelectionListProps, "premiumIconSource"> {
@@ -23,13 +25,25 @@ export const RuleSelectionListDrawer: React.FC<RuleSelectionListDrawerProps> = (
   callback = () => {},
   onRuleItemClick = () => {},
 }) => {
+  const location = useLocation();
+  const navigate = useNavigate();
+  const isRedirectFromCreateRulesRoute = useIsRedirectFromCreateRulesRoute();
+
   return (
     <>
       <Drawer
         title="Create new rule"
         width={360}
         open={open}
-        onClose={onClose}
+        afterOpenChange={(isOpen) => {
+          if (!isOpen && isRedirectFromCreateRulesRoute) {
+            // Clear location.state if it was opened due to redirect
+            navigate(location.pathname, {});
+          }
+        }}
+        onClose={(e) => {
+          onClose();
+        }}
         placement={drawerPlacement}
         className="rule-selection-list-drawer"
         destroyOnClose
