+++ app/src/routes/onboardingRoutes.tsx
@@ -8,6 +8,14 @@ export const onboardingRoutes: RouteObject[] = [
     element: <Navigate to={PATHS.RULES.MY_RULES.ABSOLUTE} />,
   },
   {
+    path: PATHS.RULES.CREATE + "*",
+    element: <Navigate to={PATHS.RULES.MY_RULES.ABSOLUTE} state={{ from: PATHS.RULES.CREATE }} />,
+  },
+  {
+    path: PATHS.GETTING_STARTED,
+    element: <Navigate to={PATHS.HOME.ABSOLUTE} state={{ from: PATHS.GETTING_STARTED }} />,
+  },
+  {
     path: PATHS.INVITE.RELATIVE,
     element: <InviteView />,
   },
